+++ fiboa_cli/datasets/es_cl.py
@@ -61,14 +61,12 @@ class ESCLConverter(ESBaseConverter):
         return new
 
     def get_urls(self):
-        if not self.variant:
-            self.variant = "2024"
-            log(f"Choosing first variant {self.variant}", "warning")
+        if not self.year:
+            self.year = "2024"
+            log(f"Choosing first year {self.year}", "warning")
         else:
-            assert self.variant in "2024 2023 2022 2021 2020 2019".split(), (
-                f"Wrong variant {self.variant}"
-            )
-        base = f"http://ftp.itacyl.es/cartografia/05_SIGPAC/{self.variant}_ETRS89/Parcelario_SIGPAC_CyL_Provincias/"
+            assert self.year in "2024 2023 2022 2021 2020 2019".split(), f"Wrong year {self.year}"
+        base = f"http://ftp.itacyl.es/cartografia/05_SIGPAC/{self.year}_ETRS89/Parcelario_SIGPAC_CyL_Provincias/"
         response = requests.get(base)
         assert response.status_code == 200, f"Error getting urls {response}\n{response.content}"
         uris = {